% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_param}
\alias{get_param}
\title{Generate Parameters for RD Analysis}
\usage{
get_param(
  outcomes,
  running,
  cutoff,
  vce = "hc1",
  est = "robust",
  order = 1,
  bandwidth = NULL,
  bin_bandwidth = NULL,
  covariate = NULL
)
}
\arguments{
\item{outcomes}{A character vector specifying the outcome variables for the RD analysis.}

\item{running}{A string indicating the running variable used in the RD analysis.}

\item{cutoff}{A numeric value specifying the cutoff point for the running variable in the RD analysis.}

\item{vce}{A string specifying the variance-covariance estimator to be used. Default is "hc1".}

\item{est}{A string indicating the estimation method to be used. Default is "robust".}

\item{order}{An integer specifying the order of the polynomial to fit in the RD design. Default is 1 (linear).}

\item{bandwidth}{A numeric specifying the bandwidth around the cutoff. If NULL, it will be calculated automatically based on the data and method. Default is NULL.}

\item{bin_bandwidth}{A numeric value used to specify the bandwidth for binning data in the graphical representation of the RD design. If NULL, it will be determined automatically. Default is NULL.}

\item{covariate}{A vector of strings representing any covariates to be included in the RD analysis. Default is NULL.}
}
\value{
A list containing the specified parameters for the RD analysis.
}
\description{
This function generates a list of parameters for regression discontinuity (RD) analysis based on user-defined inputs. It is primarily used to set up the necessary parameters before conducting an RD analysis.
}
\examples{
params <- get_param(
  outcomes = c("test_score", "graduation_rate"),
  running = "age",
  cutoff = 18,
  vce = "hc1",
  est = "robust",
  order = 2,
  bandwidth = 5,
  bin_bandwidth = 0.5,
  covariate = c("gender", "income")
)

}
