% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rd_param}
\alias{get_rd_param}
\title{Generate Parameters for RD Analysis}
\usage{
get_rd_param(
  outcomes,
  running,
  cutoff,
  donut = -1,
  bandwidth = NULL,
  covs = NULL,
  cluster = NULL,
  est = "robust",
  bin_bandidth_ratio = 1,
  est_param = list(vce = "hc1"),
  plot_param = list(kernel = "triangular", p = 1)
)
}
\arguments{
\item{outcomes}{A character vector specifying the outcome variables for the RD analysis.}

\item{running}{A string indicating the running variable used in the RD analysis.}

\item{cutoff}{A numeric value specifying the cutoff point for the running variable in the RD analysis.}

\item{donut}{A numeric value specifying the size of the donut hole, default is -1}

\item{bandwidth}{A numeric value specifying the bandwidth for data used, default is "NULL"}

\item{covs}{the vector of column names of variable used as covariates, default is NULL}

\item{cluster}{the name of the column name of variable used for clustering standard error, default is NULL}

\item{est}{A string indicating the estimator to be used. Default is "robust".}

\item{bin_bandidth_ratio}{A numeric to adjust the size of the bin relative to the local regression line, NULL value just show all data, default is 1}

\item{est_param}{A list of parameters to be passed to rdrobust, default is list(vce = "hc1")}

\item{plot_param}{A list of parameters to be passed to rdplot, default is list(kernel = "triangular", p = 1)}
}
\value{
A list containing the specified parameters for the RD analysis.
}
\description{
Generates a list of parameters for regression discontinuity (RD) analysis based on user-defined inputs.
It is primarily used to set up the necessary parameters before conducting an RD analysis.
}
