% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot,easyrd_result-method}
\alias{plot,easyrd_result-method}
\title{Plot the result from easyrd}
\usage{
\S4method{plot}{easyrd_result}(x)
}
\arguments{
\item{x}{A easyrd_result object}
}
\description{
Plot the result from easyrd
}
