% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rd.R
\name{plot_alt_rd}
\alias{plot_alt_rd}
\title{Draw Alt RD}
\usage{
plot_alt_rd(dt)
}
\arguments{
\item{dt}{a data.table from result$estimate}
}
\value{
ggplot
}
\description{
draw the RD plot from the RD source with alternative specification
}
