% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rd.R
\name{plot_rd}
\alias{plot_rd}
\title{Draw RD}
\usage{
plot_rd(ps)
}
\arguments{
\item{ps}{plot source or easy rd plot object or a data.table from result$plot_source}
}
\value{
ggplot
}
\description{
draw the RD plot from the RD source
}
