% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary,easyrd_result-method}
\alias{summary,easyrd_result-method}
\title{Summarize the result from easyrd}
\usage{
\S4method{summary}{easyrd_result}(object)
}
\arguments{
\item{object}{A easyrd_result object}
}
\description{
Summarize the result from easyrd
}
