% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplerd.R
\name{simplerd}
\alias{simplerd}
\title{Simple RD Analysis}
\usage{
simplerd(
  data,
  p,
  alt_type = NULL,
  values = c(),
  result_type = c("estimate", "plot"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.table containing the data for RD analysis.}

\item{p}{A list containing parameters for the RD analysis, typically including outcomes and other necessary information.}

\item{alt_type}{A string that specifies the type of alternative analysis to be conducted. Valid types are NULL (defaults to "main"), "subsample", and "donut". This argument determines how the data is manipulated before analysis.}

\item{values}{A vector of values corresponding to the alt_type specification. These are used to modify the dataset or parameters based on the type of RD analysis being conducted.}

\item{result_type}{A character vector indicating the types of results to return. The default is c("estimate", "plot"), meaning both estimates and plots will be generated. Can also include "plot_source" for the plot data.}

\item{verbose}{A boolean flag indicating whether to print messages for each result produced. Defaults to FALSE}

\item{...}{Additional arguments passed to the estimation function est_rd.}
}
\value{
A list containing the analysis results. This includes 'estimate' (data.table of estimates), 'plot_source' (list of data for plots), and 'plot' (list of generated plots), depending on the specified result_type.
}
\description{
This function simplifies regression discontinuity (RD) analysis results by providing estimates and plots based on the specified alternative specifications and values.
It allows for different types of RD analyses including main, subsample, and donut approaches.
}
